requirements.txt
beautifulsoup4
discord.py
python-dotenv
Requests
coloredlogs
google-api-python-client
